# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Femwell Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
...
  /products/{productId}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a product review
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── CART ────────────────────────────────────────────────────────────────────
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared
  /cart/{itemId}:
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ─── ORDERS ──────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  # ─── SUBSCRIPTIONS ───────────────────────────────────────────────────────────
...
  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/posts/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, description, price, category, imageUrl, featured, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        inStock:
          type: boolean
        contents:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
    ProductInput:
      type: object
      required: [name, description, price, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        contents:
          type: array
          items:
            type: string

    # ─── REVIEW ─────────────────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, productId, userName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating, comment, userName]
      properties:
        rating:
          type: integer
        comment:
          type: string
        userName:
          type: string

    # ─── CART ────────────────────────────────────────────────────────────────────
    Cart:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        promoCode:
          type: ["string", "null"]
        discount:
          type: ["number", "null"]
    CartItem:
      type: object
      required: [id, productId, productName, price, quantity, imageUrl]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer

    # ─── ORDER ───────────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, userId, items, total, status, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        paymentMethod:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        createdAt:
          type: string
        estimatedDelivery:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [paymentMethod, deliveryAddress]
      properties:
        paymentMethod:
          type: string
          enum: [mtn_momo, telecel_cash, airteltigo, visa, mastercard]
        deliveryAddress:
          type: string
        promoCode:
          type: string

    # ─── SUBSCRIPTION ─────────────────────────────────────────────────────────────
...
    BlogPost:
      type: object
      required: [id, title, excerpt, content, category, author, imageUrl, publishedAt, featured]
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        imageUrl:
          type: string
        publishedAt:
          type: string
        featured:
          type: boolean
        readTime:
          type: integer
        tags:
          type: array
          items:
            type: string
    BlogPostInput:
      type: object
      required: [title, excerpt, content, category, author]
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
...