# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Femwell Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  items: CartItem[];
  total: number;
  status: OrderStatus;
  paymentMethod: string;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  trackingNumber?: string | null;
  createdAt: string;
  /** @nullable */
  estimatedDelivery?: string | null;
}